+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.8.0] - 2019-03-02
 ### Added
-- New property Changelog.tagNameBuilder to customize how the tag names are generated - [#10]
+- New property `Changelog.tagNameBuilder` to customize how the tag names are generated - [#10]
+- New `Release.setDate()` function - [#9]
 
 ### Fixed
 - Update `yargs-parser` dependency to v13.0.0
@@ -112,6 +113,7 @@ First version
 [#5]: https://github.com/oscarotero/keep-a-changelog/issues/5
 [#6]: https://github.com/oscarotero/keep-a-changelog/issues/6
 [#7]: https://github.com/oscarotero/keep-a-changelog/issues/7
+[#9]: https://github.com/oscarotero/keep-a-changelog/issues/9
 [#10]: https://github.com/oscarotero/keep-a-changelog/issues/10
 
 [0.8.0]: https://github.com/oscarotero/keep-a-changelog/compare/v0.7.0...v0.8.0