+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.8.1] - 2019-05-27
+### Fixed
+- Support for more than one unreleased version
+
 ## [0.8.0] - 2019-03-02
 ### Added
 - New property `Changelog.tagNameBuilder` to customize how the tag names are generated - [#10]
@@ -116,6 +120,7 @@ First version
 [#9]: https://github.com/oscarotero/keep-a-changelog/issues/9
 [#10]: https://github.com/oscarotero/keep-a-changelog/issues/10
 
+[0.8.1]: https://github.com/oscarotero/keep-a-changelog/compare/v0.8.0...v0.8.1
 [0.8.0]: https://github.com/oscarotero/keep-a-changelog/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/oscarotero/keep-a-changelog/compare/v0.6.7...v0.7.0
 [0.6.7]: https://github.com/oscarotero/keep-a-changelog/compare/v0.6.6...v0.6.7